Seperate config endpoints
